pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "infrash"
version = "0.1.2"
description = "Inteligentny runner do rozwiązywania problemów infrastrukturalnych"
readme = "README.md"
authors = [
    {name = "Twoje Imię", email = "twoj.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Installation/Setup",
    "Topic :: System :: Systems Administration",
]
keywords = ["infrastructure", "deployment", "automation", "diagnostics", "devops"]
dependencies = [
    "click>=8.0.0",
    "GitPython>=3.1.0",
    "requests>=2.25.0",
    "rich>=10.0.0",
    "pyyaml>=6.0",
    "psutil>=5.9.0",
    "packaging>=21.0",
]
requires-python = ">=3.8"

[project.urls]
"Homepage" = "https://github.com/username/infrash"
"Bug Tracker" = "https://github.com/username/infrash/issues"

[project.scripts]
infrash = "infrash.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]